ml-multidigit/requirements.txt
flask
tensorflow
pillow
numpy
flask-cors
opencv-python
